Plant Sense AI/requirements.txt
streamlit==1.36.0
flask==3.0.3
pillow==10.4.0
python-dotenv==1.2.1

